DEV: Refactor wizard preview components to use inheritence
The previous `createPreviewComponent` implementation was problematic for template colocation. We can achieve the same result using normal component class inheritance.
