+++ app/assets/javascripts/wizard/addon/components/styling-preview.hbs
