+++ app/assets/javascripts/wizard/addon/components/wizard-preview-base.hbs
